# Conflicts:
#	patches/tModLoader/Terraria/Main.cs.patch
#	patches/tModLoader/Terraria/ModLoader/Core/MdbManager.cs
#	patches/tModLoader/Terraria/ModLoader/Engine/GoGVerifier.cs
#	patches/tModLoader/Terraria/ModLoader/ModLoader.cs
#	patches/tModLoader/Terraria/ModLoader/UI/UICreateMod.cs
#	patches/tModLoader/Terraria/ModLoader/UI/UIDeveloperModeHelp.cs

+++ patches/tModLoader/Terraria/ModLoader/UI/UIDeveloperModeHelp.cs
@@ -148,6 +148,34 @@ namespace Terraria.ModLoader.UI
 			Process.Start("https://dotnet.microsoft.com/download/dotnet-core/current/runtime");
 		}
 
+		private void DownloadMono() {
+			Process.Start("https://www.mono-project.com/download/stable/");
+		}
+
+		private static bool monoStartScriptsUpdated;
+		private void UpdateMonoStartScripts() {
+			try {
+				// upgrade start scripts to system mono
+				foreach (var monoPath in new[] { "tModLoader", "tModLoaderServer" })
+					File.Copy("tModLoader-mono", monoPath, true);
+
+/*				// vanilla start scripts need to be upgraded to copy back the sys/ folder
+				var kickPaths = new List<string> { "TerrariaServer" };
+				if (!File.ReadAllText("Terraria").Contains("forwarder"))
+					kickPaths.Add("Terraria");
+
+				foreach (var kickPath in kickPaths)
+					File.Copy("tModLoader-kick", kickPath, true);*/
+
+				monoStartScriptsUpdated = true;
+				_updateRequired = true;
+			}
+			catch (Exception e) {
+				Logging.tML.Error(e);
+				Interface.errorMessage.Show("Failed to copy mono start scripts\n" + e, Interface.developerModeHelpID);
+			}
+		}
+
 		private void DownloadModCompile() {
 			SoundEngine.PlaySound(SoundID.MenuOpen);
 			// download the ModCompile for the platform we don't have
@@ -163,13 +191,17 @@ namespace Terraria.ModLoader.UI
 					var currentEXEFilename = Process.GetCurrentProcess().ProcessName;
 					string originalXMLFile = Path.Combine(ModCompile.modCompileDir, "tModLoader.xml");
 					string correctXMLFile = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), $"{currentEXEFilename}.xml");
-					File.Copy(originalXMLFile, correctXMLFile, true);
-					File.Delete(originalXMLFile);
+					if (originalXMLFile != correctXMLFile) {
+						File.Copy(originalXMLFile, correctXMLFile, true);
+						File.Delete(originalXMLFile);
+					}
 					string originalPDBFilename = ReLogic.OS.Platform.IsWindows ? "tModLoader.pdb" : (ReLogic.OS.Platform.IsLinux ? "tModLoader_Linux.pdb" : "tModLoader_Mac.pdb");
 					string originalPDBFile = Path.Combine(ModCompile.modCompileDir, originalPDBFilename);
 					string correctPDBFile = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), $"{currentEXEFilename}.pdb");
-					File.Copy(originalPDBFile, correctPDBFile, true);
-					File.Delete(originalPDBFile);
+					if (originalPDBFile != correctPDBFile) {
+						File.Copy(originalPDBFile, correctPDBFile, true);
+						File.Delete(originalPDBFile);
+					}
 					// Move the remaining XML documentation files to references folder.
 					if (!Directory.Exists(ModCompile.modReferencesPath))
 						Directory.CreateDirectory(ModCompile.modReferencesPath);
